Minor bug fixes and additions to Interrupt+NMEA
Fixed interrupt_nmea LED error code by implementing SPI.end() in the
error function.

Fixed CorrectJumps.py issue with first data after connecting.
Before, the data was deemed good/bad based on whether there was a fix.
However, the PPS typically comes in ~400 ms before the nmea sentences
(only when there is a fix). If it goes from no fix -> fix in this
window, the PPS measurement is outdated but the data was still deemed as
"good".
Now there is a consistency check on the PPS-Serial time and the data is
discarded if it is too long.

Also more data files