Added Garmin files
Pretty much copied over all of the Interrupt+NMEA Analysis files (not all have been updated to work!!)

Added a few results

Added arduino files to
-> collect data
-> set the GPS properties (I haven't successfully set which sentences to display though)

Note:
The output files from measurements are in the same format as the Interrupt+NMEA outputs, but the GGA and RMC are switched! This means that some fields in the Analysis programs have to be adjusted.