No real changes to CorrectData.py
CorrectJumps.py has been updated to be in line with Garmin version. No huge change in performance; instead of subtracting an integer number of 1000 ms from jumps, the jumps are now reset to 1000 ms.
This works better for big jumps because we don't get "amplification" of the difference over the course of many seconds crammed into a single second. The bad part is that the distribution of PPS-PPS times has more 1000 ms than it should.
Both the previous and current methods preserve PPS-serial times though.